# backend/swagger/v1/swagger.json
{
    "openapi": "3.0.1",
    "info": {
      "title": "Your API",
      "version": "v1"
    },
    "paths": {
...
      "/api/v{version}/datasets": {
        "post": {
          "tags": [
            "Datasets"
          ],
          "parameters": [
            {
              "name": "version",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string"
              }
            }
          ],
          "requestBody": {
            "content": {
              "multipart/form-data": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "Title": {
                      "type": "string"
                    },
                    "Description": {
                      "type": "string"
                    },
                    "Archive": {
                      "type": "string",
                      "format": "binary"
                    }
                  }
                },
                "encoding": {
                  "Title": {
                    "style": "form"
                  },
                  "Description": {
                    "style": "form"
                  },
                  "Archive": {
                    "style": "form"
                  }
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "OK"
            }
          }
        },
        "get": {
          "tags": [
            "Datasets"
          ],
          "parameters": [
            {
              "name": "version",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "OK"
            }
          }
        }
      },
      "/api/v{version}/datasets/{id}": {
        "get": {
          "tags": [
            "Datasets"
          ],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            },
            {
              "name": "version",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "OK"
            }
          }
        },
        "patch": {
          "tags": [
            "Datasets"
          ],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            },
            {
              "name": "version",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "OK"
            }
          }
        },
        "delete": {
          "tags": [
            "Datasets"
          ],
          "parameters": [
            {
              "name": "id",
              "in": "path",
              "required": true,
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            },
            {
              "name": "version",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "OK"
            }
          }
        }
      },
...
      "/api/v{version}/reports": {
        "post": {
          "tags": [
            "Reports"
          ],
          "parameters": [
            {
              "name": "version",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string"
              }
            }
          ],
          "requestBody": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportDTOModel"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportDTOModel"
                }
              },
              "application/*+json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportDTOModel"
                }
              }
            }
          },
          "responses": {
            "200": {
              "description": "OK"
            }
          }
        },
        "get": {
          "tags": [
            "Reports"
          ],
          "parameters": [
            {
              "name": "version",
              "in": "path",
              "required": true,
              "schema": {
                "type": "string"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "OK"
            }
          }
        }
      },
...
    },
    "components": {
      "schemas": {
...
        "ReportDTOModel": {
          "type": "object",
          "properties": {
            "id": {
              "type": "integer",
              "format": "int32"
            },
            "markedId": {
              "type": "integer",
              "format": "int32"
            },
            "creatorId": {
              "type": "integer",
              "format": "int32"
            },
            "comment": {
              "type": "string",
              "nullable": true
            },
            "loadDatetime": {
              "type": "string",
              "format": "date-time"
            }
          },
          "additionalProperties": false
        },
...
      }
    }
  }